Move BeaconDetails components into their own files - AirBnb way :heart:
